chore: switch to in-memory repo as on-disk/level one is slow
