+++ package.json
@@ -5,8 +5,7 @@
   "leadMaintainer": "Dirk McCormick <dirk@protocol.ai>",
   "main": "src/index.js",
   "browser": {
-    "./test/utils/create-libp2p-node": false,
-    "./test/utils/create-temp-repo-nodejs.js": "./test/utils/create-temp-repo-browser.js"
+    "./test/utils/create-libp2p-node": false
   },
   "types": "dist/src/index.d.ts",
   "typesVersions": {
@@ -71,6 +70,7 @@
     "assert": "^2.0.0",
     "benchmark": "^2.1.4",
     "delay": "^5.0.0",
+    "interface-datastore": "^4.0.1",
     "ipfs-repo": "^9.1.0",
     "ipfs-utils": "^6.0.1",
     "iso-random-stream": "^2.0.0",
