+++ test/bitswap.js
@@ -8,7 +8,7 @@ const pWaitFor = require('p-wait-for')
 
 const Bitswap = require('../src')
 
-const createTempRepo = require('./utils/create-temp-repo-nodejs')
+const createTempRepo = require('./utils/create-temp-repo')
 const createLibp2pNode = require('./utils/create-libp2p-node')
 const makeBlock = require('./utils/make-block')
 const orderedFinish = require('./utils/helpers').orderedFinish
