+++ test/utils/create-bitswap.js
@@ -1,7 +1,7 @@
 'use strict'
 
 const Bitswap = require('../../src')
-const createTempRepo = require('./create-temp-repo-nodejs')
+const createTempRepo = require('./create-temp-repo')
 const createLibp2pNode = require('./create-libp2p-node')
 
 module.exports = async () => {
